Storybook: Add back button & action
Repurpose & rename the existing signal, emitted when a quest is
selected, with a null quest indicating that the user chose to go back.

Extract the focus-setting logic to a function, rather than burying it in
the loop. This will be used when re-displaying the storybook, as well as
handling the case where no quests are found – this would indicate an
error on our part but it's not so hard to focus something sensible in
that case.